Move the test case to main testfile
