+++ clang-tools-extra/test/clang-tidy/checkers/cppcoreguidelines/pro-type-member-init-no-crash.cpp
@@ -13,21 +13,3 @@ template <typename T> class NoCrash {
     template <typename U> B(U u) {}
   };
 };
-
-namespace gh192510 {
-  template<typename T>
-  struct C {
-
-  };
-
-  struct Base {  
-    int x;
-  };
-
-  template<typename T>
-  class X: public Base {
-    using INT = C<T>;
-
-    X(INT i) : INT(i) {} // no crash
-  };
-}
